iio: adc: Fix reversed mode setting logic
The mode setting logic in ad7816_store_mode was reversed due to
incorrect handling of the strcmp return value. strcmp returns 0 on
match, so the `if (strcmp(buf, "full"))` block executed when the
input was not "full".

This resulted in "full" setting the mode to AD7816_PD (power-save) and
other inputs setting it to AD7816_FULL.

Fix this by using sysfs_streq to correctly check for "full" and
"power-save", mapping them to AD7816_FULL and AD7816_PD respectively.

Signed-off-by: Gabriel Shahrouzi <gshahrouzi@gmail.com>